Make kazuki optional, as it is not used at runtime anyway
