+++ plugins/capabilities/nexus-capabilities-plugin/pom.xml
@@ -88,6 +88,7 @@
       <artifactId>nexus-kazuki-plugin</artifactId>
       <type>${nexus-plugin.type}</type>
       <scope>provided</scope>
+      <optional>true</optional>
     </dependency>
 
     <dependency>
