+++ plugins/capabilities/nexus-capabilities-plugin/src/main/java/org/sonatype/nexus/plugins/capabilities/internal/CapabilityRegistryBooter.java
@@ -16,7 +16,7 @@ import javax.inject.Inject;
 import javax.inject.Named;
 
 import org.sonatype.nexus.plugins.capabilities.internal.storage.CapabilityStorage;
-import org.sonatype.nexus.plugins.capabilities.internal.storage.KazukiCapabilityStorageConverter;
+import org.sonatype.nexus.plugins.capabilities.internal.storage.CapabilityStorageConverter;
 import org.sonatype.nexus.proxy.events.NexusInitializedEvent;
 import org.sonatype.nexus.proxy.events.NexusStoppingEvent;
 import org.sonatype.sisu.goodies.eventbus.EventBus;
@@ -40,17 +40,17 @@ public class CapabilityRegistryBooter
 
   private final Provider<CapabilityStorage> capabilityStorage;
 
-  private final Provider<KazukiCapabilityStorageConverter> storageConverter;
+  private final Provider<CapabilityStorageConverter> storageConverter;
 
   @Inject
   public CapabilityRegistryBooter(final EventBus eventBus,
                                   final Provider<DefaultCapabilityRegistry> capabilityRegistry,
                                   final Provider<CapabilityStorage> capabilityStorage,
-                                  final Provider<KazukiCapabilityStorageConverter> storageConverter)
+                                  final Provider<CapabilityStorageConverter> storageConverter)
   {
     this.capabilityRegistry = capabilityRegistry;
     this.capabilityStorage = checkNotNull(capabilityStorage);
-    this.storageConverter = checkNotNull(storageConverter);
+    this.storageConverter = storageConverter;
     checkNotNull(eventBus).register(this);
   }
 
@@ -59,7 +59,10 @@ public class CapabilityRegistryBooter
     try {
       capabilityStorage.get().start();
 
-      storageConverter.get().maybeConvert();
+      CapabilityStorageConverter converter = storageConverter.get();
+      if (converter != null) {
+        converter.maybeConvert();
+      }
 
       capabilityRegistry.get().load();
     }
