+++ plugins/capabilities/nexus-capabilities-plugin/src/main/java/org/sonatype/nexus/plugins/capabilities/internal/guice/GuiceModule.java
@@ -21,6 +21,7 @@ import org.sonatype.nexus.configuration.application.ApplicationDirectories;
 import org.sonatype.nexus.plugins.capabilities.internal.ActivationConditionHandlerFactory;
 import org.sonatype.nexus.plugins.capabilities.internal.ValidityConditionHandlerFactory;
 import org.sonatype.nexus.plugins.capabilities.internal.storage.CapabilityStorage;
+import org.sonatype.nexus.plugins.capabilities.internal.storage.KazukiPresenceChecker;
 import org.sonatype.nexus.plugins.capabilities.internal.storage.XmlCapabilityStorage;
 import org.sonatype.nexus.plugins.capabilities.internal.validator.ValidatorFactory;
 
@@ -55,71 +56,8 @@ public class GuiceModule
     install(new FactoryModuleBuilder().build(ValidityConditionHandlerFactory.class));
     install(new FactoryModuleBuilder().build(ValidatorFactory.class));
 
-    install(new LifecycleModule("nexuscapability"));
-
-    bind(JdbiDataSourceConfiguration.class).annotatedWith(Names.named("nexuscapability"))
-        .toProvider(JdbiConfigurationProvider.class).in(Scopes.SINGLETON);
-
-    install(new EasyKeyValueStoreModule("nexuscapability", null)
-        .withSequenceConfig(getSequenceServiceConfiguration())
-        .withKeyValueStoreConfig(getKeyValueStoreConfiguration())
-    );
-  }
-
-  private SequenceServiceConfiguration getSequenceServiceConfiguration() {
-    SequenceServiceConfiguration.Builder builder = new SequenceServiceConfiguration.Builder();
-
-    builder.withDbType("h2");
-    builder.withGroupName("nexus");
-    builder.withStoreName("capability");
-    builder.withStrictTypeCreation(true);
-
-    return builder.build();
-  }
-
-  private KeyValueStoreConfiguration getKeyValueStoreConfiguration() {
-    KeyValueStoreConfiguration.Builder builder = new KeyValueStoreConfiguration.Builder();
-
-    builder.withDbType("h2");
-    builder.withGroupName("nexus");
-    builder.withStoreName("capability");
-    builder.withPartitionName("default");
-    builder.withPartitionSize(100_000L);
-    builder.withStrictTypeCreation(true);
-    builder.withDataType(CAPABILITY_SCHEMA);
-
-    return builder.build();
-  }
-
-
-  // TODO: Extract helper for jdbi config, as the location for databases will be normalized
-
-  private static class JdbiConfigurationProvider
-      implements Provider<JdbiDataSourceConfiguration>
-  {
-    private final ApplicationDirectories directories;
-
-    @Inject
-    public JdbiConfigurationProvider(final ApplicationDirectories directories) {
-      this.directories = checkNotNull(directories);
-    }
-
-    @Override
-    public JdbiDataSourceConfiguration get() {
-      JdbiDataSourceConfiguration.Builder builder = new JdbiDataSourceConfiguration.Builder();
-
-      builder.withJdbcDriver("org.h2.Driver");
-
-      File dir = directories.getWorkDirectory("db/capabilities", false);
-      File file = new File(dir, dir.getName());
-      builder.withJdbcUrl("jdbc:h2:" + file.getAbsolutePath());
-
-      builder.withJdbcUser("root");
-      builder.withJdbcPassword("not_really_used");
-      builder.withPoolMinConnections(25);
-      builder.withPoolMaxConnections(25);
-
-      return builder.build();
+    if (KazukiPresenceChecker.PRESENT) {
+      install(new KazukiModule());
     }
   }
 }
