+++ plugins/yum/nexus-yum-repository-plugin/pom.xml
@@ -69,6 +69,7 @@
       <artifactId>nexus-kazuki-plugin</artifactId>
       <type>${nexus-plugin.type}</type>
       <scope>provided</scope>
+      <optional>true</optional>
     </dependency>
 
     <dependency>
